refactor

+++ mask/mask.go
@@ -28,7 +28,7 @@ type (
 		Alias    string `yaml:"alias,omitempty"`
 		Required bool   `yaml:"required,omitempty"`
 		Type     string `yaml:"type,omitempty"` // password/date - 2017-07-21/date_time - 2017-07-21T17:32:28Z/date_epoch - 1631399887
-		Index    int    `yaml:"index,omitempty"`
+		Index    int64    `yaml:"index,omitempty"`
 		IsMulti  bool   `yaml:"is_multi,omitempty"` // is this a multi-select field
 	}
 )
